fileSD: escape values passed to glob.glob
Since glob treats *, ?, and values between [] as special characters, we
need to first escape the string before passing it to glob.
Since this patch escapes the string in such cases, it also fixes the bug
where an IPV6 address is treated as a character range.
For example: "[201::1]:/path" will now be first escaped to
"[[]201::1]:/path" and only then passed to glob.

Change-Id: Id13b5adadc8e2a2be0d32105cb22b35544db1fad
Bug-Url: https://bugzilla.redhat.com/1302358
Signed-off-by: Idan Shaby <ishaby@redhat.com>