executor: Don't start thread in __init__
Starting a thread in __init__ is easy but always a bad idea, in
particular when starting the thread in the middle of __init__. Here it
causes a deadlock in the executor tests, if we add logging in
concurrent.thread() when a thread is started.

The failure flow is this:

1. Worker starts thread in the middle of __init__, before self._task is
   initialized.
2. Thread function wrapper logs start message with the target function,
   Worker._run.
3. Worker._run function __repr__ invokes Worker.__repr__
4. Worker.__repr__ tries to access non-existing self._task and raises
   AttributeError.
5. Thread fails, never calling Barrier.wait()
6. Test deadlocks waiting on the barrier

Fixed by extracting Worker.start() method, and calling when creating a
new worker, *after* the worker is completely constructed.

Change-Id: I07e807da06adfeaad6213ddf2cac8a34d046b12c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>