virt: Update memory devices correctly
The current algorithm of assigning aliases and addresses to memory devices in
Memory.update_device_info works basically as follows: Take the first memory
device from domain XML and match it with all unassigned memory devices.  In the
result all the memory devices get (incorrectly) the same alias and address.

This patch fixes two problems:

- It doesn't assign already assigned memory devices from domain XML.
- It checks memory node and size before considering the devices as matching.

The same problem is apparently present in other device updates as well.  Since
we may not come to a clear agreement how to generalize the fix, we fix just
Memory for now and postpone fixing the other devices.

Bug-Url: https://bugzilla.redhat.com/1228543
Change-Id: I0d836c707be800565d3a80a819daa8bbfb69018d
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>