net: Introduce add/remove sourceroute api
sourceroutes have been configured by the existing configurators, using
the sourceroute module to generate the correct configuration details.

There are users of sourceroute that do not really need the
configurators (dhclient_monitor), but had to use it as there was no
other option to add or remove a sourceroute.

This patch introduces the ability to add sourceroutes and remove them
through the sourceroute module, without the need to pass through the
configurators.
This step will allow supporting sourceroute with OVS and follows the
'driver' pattern, where each component has its own drivers.

The add/remove functionality is using the ip.rule and ip.route modules.

Change-Id: I70f1755734e3de1d7cdc26cc1c7bf755efcad1f2
Signed-off-by: Edward Haas <edwardh@redhat.com>