tests: Handle errors when qemu chain isn't valid
When qemu chain isn't valid, e.g. volume format isn't as expected (could happen
as a result of missing volumes format when running qemu rebase), qemu-io fails
to read data from the chain, but doesn't raise an exception (!), return code is
0, out is empty *but* err contains a description of the error. In this patch,
we check the returned err and raise exception if it is not empty.

Change-Id: I39dd86eaf9f118c33738e6c8b357ed4e113910f5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Ala Hino <ahino@redhat.com>