tests: bootstrap the virt subdirectory
Start to move all the virt tests into its own subdirectory.

WARNING: up until the move of real virt code under lib/ is completed,
we are forced to use the ugly tests/virttests name in place of the much
nicer tests/virt to avoid to confuse the python importer during the
test run.

Change-Id: Id8104853c61aae17a79847da01534e50eb1af1a0
Signed-off-by: Francesco Romani <fromani@redhat.com>