Cold Merge: Support merge of shared base parent
When the VM is cloned from a template and the storage allocation is
thin, the root volume of the VM is a shared volume. As shared volumes
are not returned when calling Image.getChain, it is impossible to rely
on that call when preparing the chain for merge. To handle this use
case, we manually add the template volume to the volumes chain if
exists.

Note: This patch is missing unit tests because storage-tests infra
doesn't activate the LVs so there it isn't possible to test volumes
prepare and teardown. This missing functionality is reported here:
https://bugzilla.redhat.com/1411103

Change-Id: Ib75da754e99efa5863108d90fdad82863fe7d7a1
Bug-Url: https://bugzilla.redhat.com/1408977
Signed-off-by: Ala Hino <ahino@redhat.com>