net: Support multiple dhclient-monitor clients through registration
dhclient-monitor has been used so far by a single client: sourceroute.

In order to support mutiple clients and decouple even further the
modules involved, an observer pattern approach has been taken.
It consists of a single publisher (dhclient-monitor) and multiple
subsribers (sourceroute for now, events later on).
A subsciber registers to the publisher, asking it to notify it when a
specific action has occured through a callback function.

Both the dhclient_monitor and the sourceroute are unaware of each other,
only at the initialization phase (currently, supervdsm_server) they are
being linked through the registration step.

Change-Id: I6883bac848da989e7cebd7df92f6db4a31117725
Signed-off-by: Edward Haas <edwardh@redhat.com>