monitor: Fix domain monitoring isolation
Previously we tried to acquire a host id both from the checker event
loop thread when a checker completes, and from the monitor thread. We
assumed that acquiring a host id is a fast operation since we are using
async=True. However, during other flows, we are acquiring the host id
using async=False. Acquiring the host id takes about 20 seconds
normally, but can take up to 3 minutes if the host was not shutdown
cleanly. This leads to blocking of the event loop thread, delaying
checking of unrelated domains.

The problematic flow is:

1. Request thread acquire host id with async=False
2. Request thread takes the SANLock._lock
3. Check thread finish and try to acquire with async=True
4. Check thread blocked on SANLock._lock until acquire finish

When this happens, we see this warning:

2017-07-11 19:13:35,546-0400 WARN  (check/loop) [storage.asyncutils]
Call <bound method DirectioChecker._check of <DirectioChecker
/dev/3997e663-0f09-4e29-b7de-c69975f734b4/metadata running
next_check=4295185.60 at 0x2fda350>> delayed by 1.42 seconds
(asyncutils:138)

Additionally, MonitorThread._acquireHostId is ia a cancellation point,
raising utils.Canceled exception when the monitor thread is stopped.
This is fine in the monitor thread, since it was designed for
cancellation, but raising a BaseException subclass in the event loop
abort the event loop. This can explain the silent death of the event
loop threads seen in the bug.

When this happens, we would see this log:

2017-07-12 03:35:16,889+0300 ERROR (check/loop) [root] FINISH thread
<Thread(check/loop, started daemon 139956156602112)> failed
(concurrent:198)
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/vdsm/common/concurrent.py", line 191, in run
    ret = func(*args, **kwargs)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 125, in run_forever
    self._run_once()
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 190, in _run_once
    handle._run()
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 394, in _run
    self._callback(*self._args)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 577, in reap
    self._complete(rc)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/check.py", line 337, in _check_completed
    self._complete(result)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/monitor.py", line 496, in _pathChecked
    raise utils.Canceled("Simulating cancellation in _acquireHostId")
Canceled: Simulating cancellation in _acquireHostId

Finally, acquiring a host id from two threads creates a race when
checking if host id should be released. I think this race is the root
cause for random failure in seen in bug 1449944, and so far this issue
was not seen with this patch.

This patch changes host id acquire so it is done only from the monitor
thread, so the checker thread can never block on clusterlock internal
lock, and would not get cancelled by mistake.

Change-Id: I1eb57634bcc11a5b11add5b17c67ba93bdaacca8
Bug-Url: https://bugzilla.redhat.com/1469175
Bug-Url: https://bugzilla.redhat.com/1449944
Signed-off-by: Nir Soffer <nsoffer@redhat.com>