monitor: Log changes in domain host id status
When acquiring a host id in the domain monitor we use async=True, so the
host id is not acquired yet when StorageDomain.acquireHostId() returns.
We find that the operation was successful on a future monitor cycle.

Log INFO message when StorageDomain.hasHostId() changes from False to
True (successful acquire), and log WARN message if the value changes
from True to False, meaning that the host id was released.

Change-Id: I120b10a81eca40ef3b23653c1bba1e6a48a54a2a
Bug-Url: https://bugzilla.redhat.com/1422508
Signed-off-by: Nir Soffer <nsoffer@redhat.com>