automation: prepare yum.conf before copying to lago guest
The host's yum.conf sets reposdir to /dev/null/, causing
localsync to be ignored which en route causes vdsm to be installed
via the nightly repository.

This patch both excludes vdsm the rest of the repositories configured
in yum.conf and removes reposdir from yum.conf (resulting in the default
destination which holds localsync) by modifying and copying
over a temporary copy of yum.conf.

Change-Id: Ia301a1750f8755f49285fb202c7ee8bc9555d1ca
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>