v2v: allow setting of QCOW2 image version during import
Since 4.1 it should be possible to use also newer version of QCOW2
images. Also this commit [1] adds option to virt-v2v to chose which
version to use in imported image.

First, this change makes use of the newly added option in virt-v2v
(if the installed version of virt-v2v supports it). Second, the API
commands for importing VMs from VMware/Xen/OVA are enhanced so that
engine can tell us which version to use.

[1] https://github.com/libguestfs/libguestfs/commit/bdaeeb4e606f3828

Bug-Url: https://bugzilla.redhat.com/1412451
Change-Id: I846e75e6292c0b249e66b08ceec6b371e9b50f22
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>