logUtils: Support log level aliases
We are documenting the shorter "WARN" log level, but the actual
code failed to support them. Use logging.getLevelName to support the log
level we added using logging.addLevelName.
The new log levels were added to vdsm schema as well.

The code was also not tested, add tests checking that it actually works.

Change-Id: I8ec977075da3228a660e1b126857ca6b1a25f3c0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Irit Goihman <igoihman@redhat.com>