vm: tests: actually set existing ioTune settings
Reading the source code in vm.py _and_ vmdevices/storage.py
reveals that the storage code was using and storing the ioTune
settings under self.specParams, not under self.conf['specParams'].

This is fine, but the test code feed the storage device with
pre-existing ioTune settings using self.conf['specParams'].

This is misleading in the best case, wrong in the worst case.
This patch actually sets the old ioTune settings using the right
field - specParams.

Doing so we discover that the code is indeed broken if it has
to update the existing ioTune settings - exactly because
the testcase never covered such scenario.

For the (short) time being, we mark the test broken until we fix
the code.

Change-Id: I5e266d0a8a5378d4e02b70289999d7eae0fb6ba9
Signed-off-by: Francesco Romani <fromani@redhat.com>