vm: storage: avoid pointless XML processing
Vm.setIoTune() used to
1. extract the current ioTune settings from the device XML
2. update the device XML - just before to recreate it

Point #2 was pointless and wasteful, and point #1 was cumbersome
and error prone.

It is responsability of the storage.Device class to return its
up-to-date ioTune settings; how it stores them, how it makes the
XML representation consistent and how it manage its state
are all implementation details which must be kept private.

This patch makes the vm.py code use the helpers introduced into
the parent patch, separating the responsabilities, reducing the
coupling, avoiding pointless (and misleading) work.

Change-Id: I2e185e1d6c28c31321ea07675903e543362eb957
Signed-off-by: Francesco Romani <fromani@redhat.com>