qemuimg: Replace commands.watchCmd with operation
qemuimg.rebase() returns now an operation object, similar to convert()
and commit(). The caller should run the returned operation. To abort the
operation, the caller can call the operation.abort() method, instead of
passing "stop" argumnet to qemuimg.rebase().

The return operation object is little different from QemuimgOperation.
We have another patch for unifying the interfaces of the operations.

Change-Id: I883d097f4f73d2154149f387bc0eb6b9c878f6ac
Signed-off-by: Nir Soffer <nsoffer@redhat.com>