v2v: log path to the import log file
When importing external VM we are storing complete output of virt-v2v
and kvm2ovirt in external log file under /var/log/vdsm/import/ however
the path (espcially the log file name) is not present in vdsm.log. This
makes it harder to find the relevant log after failed import.

This patch fixes this.

Change-Id: Iadf1008c392a0f89457600ac2bfab7d002ea23bc
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>