sslutils: _compare_names: fix arg names and add tests
_compare_names needs to return True if the TCP stream originates from a
source address that matches the certificate presented by the client.
This patch makes the method args clearer. Since the method does not need
to use `self`, it is declared a @staticmethod. Tests are added to
document its expected behavior and make sure the behavior is maintained
in a following patch.

Bug-Url: https://bugzilla.redhat.com/1412550
Change-Id: I7faeaaee728da4f97aaa05fddf3ab2e418cb2c40
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>