spec: Update libvirt requirement on EL7
Live snapshot is broken when using a VM lease, causing the snapshot to
fail. This breaks live snapshot and live storage migration when using a
VM lease.

The issue was fixed in libvirt-2.0.0-10.el7_3.4, we require now this
version.

A fix for Fedora is not available yet, this is tracked in
https://bugzilla.redhat.com/1415488.

Change-Id: I2d5211617c53861376969572d96e04a66a127074
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>