getVGInfo - updated pv information
When executing getVGInfo we aren't guaranteed that the cache data
is up to date as the cache isn't synchornized between the dc hosts
while getVGInfo is a HSM operation (and therefore can be executed on any
of the hosts).
This patch adds refresh of the vg and its pvs when executing getVGInfo.

Change-Id: I0af62d8d3cd361e813f06a733678586baada1e27
Bug-Url: https://bugzilla.redhat.com/1414862
Signed-off-by: Liron Aravot <laravot@redhat.com>