net: Introduce bond properties and refactoring options handling
Bond properties are introduced, acting as a superset of bond options.
Bond options are properties that affect bond operation and can be
modified, where the properties include the options and other
informative details (read only ones).

The existing options handling has been adjusted to fit the new
introduced properties with an aim to simplify and remove duplication.

Change-Id: I7f9fe408dabb7009e07cc487d04f32b12b0322e6
Signed-off-by: Edward Haas <edwardh@redhat.com>