net: modify VF MAC addresses on creation to a valid address
When a guest releases an SR-IOV virtual function, the virtual function's
previous MAC address is being restored in host. There is currently an
unresolved bug in igb driver's where it fails to reallocate the virtual
function back to the host as it doesn't accept the factory (00:00:00:00:00:00)
MAC address.

Until the issue is resolved on the driver level, this patch implements
a hack modifying the VF's MAC addresses on creation to valid addressess that
will be accepted upon reallocation.

Change-Id: Ib740d45206d41b330fd1b33f31765050cadbd679
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1410076
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>