Upgrade Mockito dependency to 2.6.5
The Mockito 1.x generation is outdated, and the beginning of the work
on oVirt 4.2 is a good time to upgrade our testing facilities.

This patch upgrades to the latest available Mockito version, 2.6.5.

Some of the benefits of this upgrade include:
1. Working with a live, maintained, library.
2. Support for Java 8 syntax.
3. Improved generics support, lending to cleaner code.
4. Improved performance

Mockito 2.x remove the org.mockito.Mockito.stub method. This patch
replaces its usage with a standard call to org.mockito.Mockito.when.

Change-Id: Icf2ab23e4ac54370a4ae7b40debcfa83e287c942
Signed-off-by: Allon Mureinik <amureini@redhat.com>