v2v: decode capacity units from OVA
When importing from OVA, we mistakenly assume that the disk capacity is
always stored in GB. This is untrue and can mean that the disk capacity
reported to engine is incorrect. This can is some cases cause the import
to fail on storage errors because engine tries to prepare image in
orders of magnitude larger than is necessary.

The correct units of the ovf:capacity attribute are stored in
ovf:capacityAllocationUnits attribute. The correct flow is to parse the
units and then properly scale the reported capacity.

Parsing routine was introduced that decodes the programatical units. The
routine is intentionaly kept simple. It does not decode all possible
values described in specification and focuses only on values we can
expect in OVA/OVF.

Bug-Url: https://bugzilla.redhat.com/1422087
Backport-To: 4.1
Change-Id: I8b865fbb4ff457a29c06a446fec9e25dd885164b
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>