storage: cleanup logs on new domain flow
When adding a new storage domain, the flow verifies that
the provided UUID does not exists.
In order to validate this, all domain types searching algorithms are
used.

During the flow, some logs were printed in ERROR or WARNING level, even in the
positive scenario.

In order to clean the logs from unneeded noise, the following changes
were made:

lvm.py - remove unneeded warning, failure to get a vg is expected
         when the vg does not exist, and we raise an exception in this case
sdc.py - remove duplicate logs about searching for an unfetched domain
sdc.py - looking for an unfetched domain is not an error, this is a normal
         condition when creating  a storage domain
sdc.py - if an unfetched domain is not found we raise, so we must not log
         the exception. This leads to logging the same exception twice.

Change-Id: I05a573c0ca6b0256b79f9934a46132c5b75b38a7
Bug-Url: https://bugzilla.redhat.com/1223538
Signed-off-by: Fred Rolland <frolland@redhat.com>