logging: remove the schema issues reporting
The schema issues are reported in case we aren't in strict mode.
However, the issue is that it spams the log.
Due to that the _log_devel was introduced, but on upgrade the log level
is the default one (DEBUG) which causes the spamming to occur.

We have set a target in 4.2 to address all the schema issues.
In the meantime (in 4.1 and 4.0) we are removing the log print here.

Change-Id: I66b7338271d44f214c8b13926dd1b29d06a9c25f
Bug-Url: https://bugzilla.redhat.com/1416315
Signed-off-by: Oved Ourfali <oourfali@redhat.com>