migration: don't require optional jobStats fields
In the migration monitor thread, we fetch the jobStats to
report the migration progress. We just log some data,
and we do logic on a small subset of this data.

We incorrectly thought some fields will always be present,
but it turns out that libvirt may omit them (after reading
the libvirt sources).

Since we use those fields only for logging, we consider them
optional too.
The fields on which we make logic will always be returned after
a succesfull jobStats() call, so no further changes are needed.

Change-Id: I09f05e36d13606020452d9b2ac634d302a0a5f8c
Bug-Url: https://bugzilla.redhat.com/1414626
Backport-To: 4.1
Backport-To: 4.0
Signed-off-by: Francesco Romani <fromani@redhat.com>