lvmlocal.conf: Remove devices filter
We want to disable lvmetad on all installations and upgrades. For this
we will add lvm configurator that install static lvmlocal.conf shipped
with vdsm.

This file cannot contain the blacklist filter, since adding such filter
when the machine has a previous filter is not safe. This cannot be part
of the vdsm-configure flow, since it will may fail vdsm installation or
upgrade even if the user is not using lvm based storage.

We should use a whitelist filter, that solve both the issue with guest
lvs, ovirt lvs and direct luns. This filter will be generated per host
and configured by a separate vdsm tool command, not part of the host
deploy flows. See https://gerrit.ovirt.org/70899.

Change-Id: Ib2b408997c0141ed6b0d6e18aafcd02640ed1840
Bug-Url: https://bugzilla.redhat.com/1398918
Signed-off-by: Nir Soffer <nsoffer@redhat.com>