spec: Bump requirement on python-cpopen
In order to support redirection of stderr to stdout we need to bump
version requirement on python-cpopen. This feature is broken in previous
versions.

The requirement was introduced by commits [1][2]:
985cbe87  v2v: Log detailed output of virt-v2v [Tomáš Golembiovský]
22d665c7  v2v: Add PipelineProc, pipeline wrapper object [Tomáš Golembiovský]

There was a previous attempt to bump the version requirement here [3]:
fa1e3804  v2v: Bump version requirement on python-cpopen [Tomáš Golembiovský]

But it was badly positioned in the first place (would still cause
breakage in runtime) and it disappeared later because of commits [4][5]:
eb803052  build: Remove build requires needed for the tests [Nir Soffer]
777f386e  build: Require python-cpopen for build [Yaniv Bronhaim]

[1] https://gerrit.ovirt.org/59834
[2] https://gerrit.ovirt.org/62094
[3] https://gerrit.ovirt.org/62093
[4] https://gerrit.ovirt.org/65500
[5] https://gerrit.ovirt.org/68986

Change-Id: I4c7a95f96c70d2c644ebb728dd471962a48e4910
Bug-Url: https://bugzilla.redhat.com/1416112
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>