migration: add boolean to control retries
We added logic to automatically retry migrations if the
destination Vdsm is over capacity.
The current logic uses the progress to learn if a retry must be
attempted.
For this purposes, it is just safer and clearer to add one new boolean
value, to be set once migration actually started.
This flag is never accessed concurrently, so we don't need a full-blown
threading.Event object.

We should never retry failed migrations without Engine's instructions,
and with the old, progress-based code, that could mistakenly happen.
With the new event-based code, we make sure this is not possible
anymore.

Change-Id: Ie45553bf3ec3db76e520d11a68c67b5b9664dc32
Related-To: https://bugzilla.redhat.com/1414626
Backport-To: 4.1
Backport-To: 4.0
Signed-off-by: Francesco Romani <fromani@redhat.com>