vdsm-tool: Add lvm configurator
This configurator masks, disables, and stops lvm2-lvmetad service and
socket, and install vdsm managed lvmlocal.conf. The installed
lvmlocal.conf is owned by vdsm and will be replaced automatically on
upgrades, unless the system administrator has marked this file as
private. Vdsm will not start now if lvmetad service or socket are not
stopped and masked.

Lvmetad was never compatible with ovirt clustered LVM solution. When we
ported ovirt to EL7, we added the use_lvmetad = 0 option in vdsm command
line lvm configuration (--config). Recently we discovered several issues
with this approach, and LVM developers advice is to disable this
service.

This patch is the first step in solving lvm auto activation issues:

1. All ovirt lvs are activated; during boot when using FC storage, and
   after connecting to storage server when using iSCSI storage.
2. Guest lvs created inside a VM using ovirt raw volume as a pv are
   activated on a host. This prevent vdsm from deactivating ovirt lvs on
   startup and when stopping a vm, and may lead to data corruption of
   the stale lv later.
3. Lvs on LUNs not used for storage domains are active on a host
4. Some LUNs are not visible to vdsm because LVM activated lvs inside a
   LUN's before multipath created the map.

This patch solves issues 1 and 2 - auto activation of ovirt logical
volumes and guest logical volumes inside ovirt raw volumes when using
iSCSI storage. The lvs are activated because by default LVM is using
event based activation via the lvmetad daemon. Disabling lvmetad
disables event based auto activation.

The patch does not solve issues 1 and 2 when using FC storage. These
issues can be fixed only using LVM filter allowing LVM to access only
the devices needed by the host. This filter must be generated on each
host, depending on the particular devices needed by this host. We cannot
do this as part of the automatic host deploy flow.  See
https://gerrit.ovirt.org/70899 for more info on generating lvm filter.

Note: We backup an existing lvmlocal.conf file, typically the example
installed by LVM without any settings, and replace it with vdsm
configuration. The system administrator should merge settings from the
backup of the file manually if needed. This is the same way we handle
multipath.conf. It would be better to merge the existing file with vdsm
configuration, we need to explore this direction later.

Note: this patch includes support for legacy ovirt-node that did not
exists in the master and 4.1 version.

Change-Id: If0e34efbcb2d8c1fac1aec3dc9e68a65957c84c2
Bug-Url: https://bugzilla.redhat.com/1398918
Signed-off-by: Nir Soffer <nsoffer@redhat.com>