preliminary support for engine's domxml
This patch introduces a basic preliminary support for getting libvirt's
domxml generated by the engine. The reason for having this temporary
code is twofold:
1. it enables people to use this new API in their development environment
in an early stage of 4.2, which increases the chance to find issues
in the generation of the XML.
2. it enables people to use this new API that is expected to simplify
the implementation of new features/changes.

The domxml is optional. If the engine sends it, it send it together with
all the stuff that the engine used to send on run/create VM. That way,
the existing code in VDSM works as it worked before and only the function
_buildDomainXML is changed: instead of building the XML from the data
in the VM dictionary, it just replace some place-holders in the given
XML and then continues with the modified XML.

Change-Id: I469fad3ca14a6b7f4675ef5c200175053f6dd4af
Signed-off-by: Arik Hadas <ahadas@redhat.com>
Signed-off-by: Francesco Romani <fromani@redhat.com>