hooks: Add localdisk hook
This hook add the ability to use fast local storage instead of shared
storage, while using shared storage for managing VM templates.

A user will create a VM normally on shared storage of any type.
To use the VM with local storage, the user will have to pin the VM to a
certain host, and enable the localdisk hook.

When starting the VM on the pinned host, the localdisk hook will copy
the VM disks from shared storage into the host local storage, and modify the
disk path to use the local copy of the disk.

The original disk may be a single volume or a chain of volumes based
on a template. The local copy is a raw preallocated volume using a LVM
logical volume on the special "ovirt-local" volume group.

Most operations are not supported on a VM using the localdisk hook.
For example, the VM cannot be migrated to another host, cannot create
snapshots, copy the local disk to other storage, etc.

An LV tag "UPDATING" is used to make sure that the copy operation
succeeded.

Here a description of the hook flow:
- If a matching local LV exists without the "UPDATING" tag:
  - activate the LV
  - update the disk path and format and return.
- If no matching local LV exists:
  - create a new LV with "UPDATING" tag.
  - copy the data from the shared storage disk/template.
  - once the copy operation is done, remove the "UPDATING" tag.
  - update the disk path and format and return.
- If a matching local LV exists with the "UPDATING":
  - delete the LV
  - create a new LV with "UPDATING" tag.
  - copy the data from the shared storage disk/template.
  - once the copy operation is done, remove the "UPDATING" tag.
  - update the disk path and format and return.

Change-Id: I5b6ad8e8ebb4cd71bff02157f28cd2be5c31ac35
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Fred Rolland <frolland@redhat.com>