net: vlan@bond config with ifcfg requires hwaddrs sync check
Bonds are expected to inherit one of its slaves mac address, but in some
cases (mainly when its slaves are detached temporary) the bond is left
with its own generated address. In such scenarios, a VLAN which may be
configured over it will inherit its temporary address and stick with it.
At the time the slaves re-attach to the bond, the VLAN interface will
not have the same address as the bond.

This has been identified as a problem when adding a network over an
existing vlan@bond NM setup.

With this patch, when the ifcfg configurator attempts to configure
a vlan over a bond, it will check post the ifup if the slave-bond-vlan
mac addresses are in sync.
In case they are not, the vlan is taken down (ifdown) and recreated.

Change-Id: I8fc13f42ec21fef1e62f326981742bd66a792429
Bug-Url: https://bugzilla.redhat.com/1414323
Signed-off-by: Edward Haas <edwardh@redhat.com>
(cherry picked from commit d2f8b7d49d069b5117da93b13d5ad997b6be0658)