amend_volume: Prepare the entire volume chain on amend.
When calling qemu-img amend on a QCOW volume based on a block storage
all the backing files should be prepared as well so their LVs will be
activated for the command to succeed.

Therefore this patch fixes the issue that was introduced in patch
https://gerrit.ovirt.org/#/c/68987
and prepare the volume by initializong the justme flag as false.

Change-Id: Id52a97b4c484d00c8913933e9a950a4a58b709c7
Bug-Url: https://bugzilla.redhat.com/1417460
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>