lib: do not use virt-prefixed logger for utilities
The @api.method code was born in the virt vertical,
but was later placed in the common area, hoping to be useful
for other verticals and used across the codebase.

However, it still incorrectly uses a virt-prefixed logger, and this
is conflicting with its stated goal: this is no more virt-specific code.

This patch replaces the 'virt' prefix with the 'vdsm' one.

Change-Id: I091fb7336db58aa38af5b01c33d6d9239afff6d4
Signed-off-by: Francesco Romani <fromani@redhat.com>