after_disk_prepare: Add new hook point
Add new after_disk_prepare hook point, called after preparing a vdsm
image. If the hook modifies the path to the image, vdsm will use the new
path.

This hook point is required for the new localdisk hook. The hook will
copy a disk from shared storage to local host storage, practically
pinning a VM disk to the host.

The hook is supported only when preparing images when starting a vm.
Other flows preparing drives are not supported.

A hook can use hooking.read_json() to get a dict with the disk
info in this format:

{
    "disk": {
        "device": "disk",
        "domainID": "17827a49-c2fd-479a-bc63-8d2a73e1d42f",
        "format": "cow",
        "imageID": "8988df9b-5d5b-445c-8150-26bbccc020dd",
        "path": "/path/to/disk",
        "volumeID": "143e2e28-b2bb-42d4-93c6-835368d3d0ae"
        ...
    }
}

See the VmDiskDevice in vdsm schema for the complete contents of the
disk info dict.

The VM uuid and custom properties are available via environment
variables.

If the hook wants to change the disk, it must write the modified disk
info using hooking.write_json(). Vdsm will use the modified disk as
returned by the hook instead of the original disk.

Here is an example hook using this hook point:

    #!/usr/bin/python2
    import os
    import sys
    import hooking

    if "hookname" not in os.environ:
        sys.exit(0)

    disk = hooking.read_json()
    disk["path"] = "/no/disk/for/you!"
    hooking.write_json(disk)

Hooks using this hook point must be installed in:

    /usr/libexec/vdsm/hooks/after_disk_prepare/

Change-Id: Ib150682c4afd7f387eabf6873b10eb5eaa858417
Bug-Url: https://bugzilla.redhat.com/1418145
Signed-off-by: Nir Soffer <nsoffer@redhat.com>