net test: Add net qos test to the new func tests.
This patch mimics testSetupNetworkOutboundQos from the legacy func tests
and adds it to the new functional tests.

It also fixes a problem in the qos assertion.

Note: the older functional test is not yet removed as it also tests the
nic qos configuration after the network is removed. For that check, an
integration test is in order (as the VDSM api does not report qos on
nics, just on networks).

Change-Id: I308de9bbce4ee6e7b2e34e419cba7b48d6c24962
Bug-Url: https://bugzilla.redhat.com/1302020
Signed-off-by: Edward Haas <edwardh@redhat.com>