lvm-conf: Handle lvmetad failed state
When configuring lvmetad, we stop the units. If the unit fail to stop
after a timeout, systemd will kill the unit, and it will move to
"failed" state instead of "stopped". This looks like a bug in systemd,
since we just asked to stop the service, but it is possible that the
service was in failed state before we stopped it, so we should support
this.

The current code will fail to start vdsm since vdsm-tool is-configured
fails, and running vdsm-tool configure does not change the service
state.

Now we check that the unit is not "active" instead of "inactive".

Change-Id: I83d4ffcca510f77ec229f2b077ec822b00d68de2
Bug-Url: https://bugzilla.redhat.com/1398918
Signed-off-by: Nir Soffer <nsoffer@redhat.com>