Freeze version of python2-crypto in check-merged
Newer version of python2-crypto break python2-paramiko which is required
by Lago.  This patch adds an EPEL snapshot that contains an older
version of python2-crypto and tries to enforce installing the older
version to make Lago work.

This patch also removes the link to EPEL in the *.repos file because
EPEL is provided natively by Standard-CI

Change-Id: I5fb6fbc84297254c87bfc5aa2a67799ce2d0f0f0
Signed-off-by: Barak Korren <bkorren@redhat.com>