virt: Make Memory device XML more complete
Memory.getXML() doesn't identify the memory device completely.  This may in
theory cause problems with memory hotunplug.  According to libvirt
documentation, complete device definition is required for hotunplug.  Although
the current definition is actually accepted, it doesn't identify the memory
device uniquely.

Assume there are two memory devices with the same size and on the same node.
One of them can't be hotunplugged for some reason, e.g. it hasn't been onlined
as movable, while the other one is fine.  However, if libvirt selects the first
device on the hotunplug request then the second device can never be
hotunplugged.

This patch adds the missing information to the Memory device XML to make it
uniquely identified and the information for libvirt complete.  Since the
additional device attributes are not available before memory hotplug, they are
inserted only if they are available.

Change-Id: Ib7037bc9b36fa3e0ead24d93cf236ed0b46aba80
Backport-To: 4.1
Bug-Url: https://bugzilla.redhat.com/1228543
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>