Revert "vdsm: remove 3.6 supported engine"
This reverts commit 7a173cc62e5b0fc79ab68198c75896e4ac751a37

I am not familiar with a specific bug showing up when Engine is 3.6 and
host is 4.y. This combination is actually very important for
hosted-engine upgrade from 3.6 to 4.1.

We must test and support 3.6 Engine and 4.1 hosts, fix specific bugs
that might pop up; not to disable this combination in its entirety.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1403846
Change-Id: Iea066a04da25e765d65f36b2e67d0034ca9f458e