virt: Fix of wrong exit code on host shutdown
In case of a host shutdown previously we did not send the ERROR
exit_code however NORMAL which was incorrect.
Due to a bug in the ovirt-engine the behavior was coincidentally
restored and the issue hasn't been uncovered during the implementation
of the host shutdown detection.
This patch will no ensure that the ERROR exit_code is sent on a detected
host shutdown.

Change-Id: I397f1e069d650b39070b38ac1c5de07f2a76a1f2
Backport-To: 4.1
Bug-Url: https://bugzilla.redhat.com/1341106
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>