virt: move vmxml.Domain to its own module
We move vmxml.Domain into the newly introduced libvirtxml.py module.
This way we isolate the utility XML processing code from the code which
uses it to build the libvirt XML. The key benefit is that with this
patch we can reuse the utility code in the upcoming metadata handling,
module, avoiding circular dependencies.

Other lesser benefits are the slightly easier future removal of the XML
building code, and the fact that 'libvirtxml.Domain' reads a bit nicer.

Change-Id: Ie2cecd05ac50c1514045219983bac79f0ff9ec09
Signed-off-by: Francesco Romani <fromani@redhat.com>