virt: storage: validate iotune params on update
The current storage.Drive code does attribute validation
and sanity check at the beginning of the getXML() method.
This is quite late in the game.

The right place to do validation is before setting the new values.
Since we just refactored the `iotune` access, we move this check
in the mutator (property setter).

Change-Id: I93fef922b2ca9f3c739d75da9288f9196af88061
Signed-off-by: Francesco Romani <fromani@redhat.com>