containers: reduce the parsing
When starting up a container, we do some preliminary parsing of the
domain XML to detect the runtime to use.
This was useful in the ealry stages, when we considered
supporting more than the docker runtime.
Nowadays we use it only for logging.

This patch eliminates this early parsing and updates the log
accordingly. This way, we save some cpu time and we make the code
a bit easier to work with for the upcoming patches.

Change-Id: I738d499fbea98aee5ecc048fe0ad9da50df4b6a9
Signed-off-by: Francesco Romani <fromani@redhat.com>