containers: require image as parameter
The container code has code to detect the parameters from
the domain XML. Among other things, it tries to be smart
and figure out the image to use, if it is not explicitely
given.

This patch starts to remove this smartness: the image to
run is expected to be given as parameter.
This will make the upcoming transition to the metadata
module easier.

We keep the fallback code which guesses the image from
the first disk drive configured. That will be removed
in a later patch.

Change-Id: Ib634733bbb2ba0c6e5242786f9b40b601f626d7e
Signed-off-by: Francesco Romani <fromani@redhat.com>