net: Replace netinfo dependency on libvirt with running config
Netinfo report has been using libvirt db for determining which networks
have been defined on the host.

As part of the effort to seperate the host networking from the VM
networking, this patch removes the dependency on libvirt db for
determining the defined networks.

This patch queries running config to determine the defined networks,
replacing the role of libvirt in this regard.

Change-Id: Ie99464833a64bb6ceea789a1fda6db6bcc5b10cc
Signed-off-by: Edward Haas <edwardh@redhat.com>