net: Persist running config accross reboots
Relocating running config from /var/run to /var/lib, making it persitent
accross reboots.

Running config represents a staging state, between the safe state
and the actual/current state (read mainly from the kernel).
As we have done so far, we assume that running config is in sync with
the actual state.

In future patches, we will aim to reduce running config content, leaving
only the information needed to identify which components are owned by
VDSM networking and possibly some additional metadata that cannot be
persisted on the device level.

Change-Id: I5fb2f53bbebb8faf57fededb15a2af8fa9dfc125
Signed-off-by: Edward Haas <edwardh@redhat.com>